DEFRMA
Now days, mobile apps are ranked by popularity. Rankings get aected by
reviews and
ratings. As the app get more positive and in favor reviews, its
popularity increases. The
popularity aects its ranking and attracts more users. Indeed, the
reviews manipulation is
becoming more common for fraud ranking of mobile apps. The imposter
often post fake
reviews with a purpose of bumping up the apps in the popularity list.
So, the necessity of
preventing fraud ranking increases. Such fake reviews must be detected
and discarded so
that real apps get to their genuine positions in the ranking. For this
purpose, the proposed
system provide a holistic view of fraud ranking by fake reviews and
provide a fake review
detection and elimination system for mobile apps. In the experiments,
the eectiveness of
the proposed system is validated, that shows the scalability of the
detection algorithm as
well as some regularity of ranking fraud activities.